Add configurable MTU to intel1g.lua
Packets which exceed the MTU are dropped on ingress and egress.
Ingress drops are performed by the NIC in hardware (note that this
commit changes the "Store Bad Package" setting from enabled to
disabled, which is the default, to achieve this behaviour).  Egress
drops are performed in software.

Ingress drops are counted by the RLEC statistics register but are
currently not exposed by the driver.  Egress drops are not counted at
all.