Improve performance of the VLAN multiplexer
The performance of the VlanMux app (apps.vlan.vlan) is improved by

  - avoiding use of pairs()
  - avoiding string composition
  - serialisation of loops to avoid branches

The forwarding strategy during de-multiplexing has been changed to
forward all packets from the trunk links and accepting drops on vlan
links which are full (instead of only forwarding as many packets as
the fullest vlan link can accept).

OTOH, packets are retained on the incoming vlan links if the trunk
link is congested.  This strategy should probably be analyzed more
closely.