intel_mp: avoid allocations and NYI in stats
The sync_stats() method uses pairs() to traverse the statistics
counters and calls a number of functions that poll hardware registers
in a manner that causes allocations on the heap to occur.  While this
happens so rarely that it is of no concern for performance, it adds
noise to the JIT activity that can make it harder to debug real
performance issues wiht -jdump. Because it is easy to avoid these
effects it seems that there is no downside in doing so.