ipv6.reassembly: perform fragment processing in separate loop
Successful reassembly of a packet ends with a call to the ctable
remove_ptr() method, which contains a loop.  That loop is frequently
leading to trace aborts and blacklisting of the main push() loop.

As a mitigation, fragments are now put on a separate queue in the main
loop and processed in a separate loop.  The level of nested loops
stays the same but apparently this leads to a better trace structure
with much fewer aborts.