ipv6.fragment: improve performance, reduce garbage
The amount of work being done in the main push() loop is reduced by
moving to-be-fragmented packets to a separate queue.  This is
beneficial if the workload is unbiased with respect to fragmentation
or if the workload shifts from mainly fragmented to mainly
unfragmented packets.

All packets needing fragmentation are processed in a separate
loop. The fragmentation itself contains another loop which might get
compiled first.  In that case, the pointers passed to
fragment_and_transmit() would require dynamic boxing on the Lua stack,
which cause GC activity.  Explicit boxing is used to avoid this.