pseudowire: move forwarding check out of the loop
Performing the check within the loop will typically yield a root trace
for the non-standard case when the pseudowire has not yet become
operational.  Once the pseudowire becomes operational, forwarding will
be performed in a sub-optimal side trace.  The new code should allow
the compiler to find a root trace for the proper forwarding loop.