Add CQE compression
CQE compression enables the hardware to reduce the number of full CQEs
getting posted to a CQ, which should provide a boost in performance
with certain traffic loads.  Essentilly, the algorithm generates a
single full CQE for a sequence of CQEs that differ only in the
checksum and the packet length and adds a compact data structure that
only stores the differing values from which the original CQEs can be
reconstructed.

The implementation was reverse-engineered from the mlx5 Linux driver
as the ConnectX4 PRM provides no documentation on the feature.

The per-queue counters "cqes" and "cqes_compressed" count the number
of uncompressed and compressed CQEs, respectively. The ratio should
give an indication of how effective the mechanism is.