Add explicit packet-rebalancing with interlinks
Interlinks require a group-wide free-list to redistribute packets
among participating processes.  A spin-lock is used to synchronize
access to that list which incurs delays during packet processing.
This added latency can cause packet drops if there is not enough
buffer space available.  As a workaround we use an additional set of
interlinks from each IPFIX worker process back to the RSS process that
originated the packets. These interlinks are connected to sinks in the
RSS process which feed the packets back to the private free-list
without having to go through the group free-list.