apps.tap.tap: use a delay to work around a bug in kernels >5.0
Newer kernels change the MAC address of a TAP interface after it was
created. The Tap app is likely to pick up the initial address which
does not match what ends up on the interface. The workaround simply
waits a while before reading the address from the kernel. This
introduces a significant startup-delay for apps that use many tap
interfaces.