apps.mellanox.connectx: overhaul async execution of firmware commands
The mechanism for executing commands in an asynchronous manner imposes
an artificial limit to the number of RSS queues that can be
configured. This is due to assigning a fixed slot in the command queue
to each asynchronous command, capping the effective number of RSS
queues by the size of the command queue because each queue requires
one command to periodically read its statistics counters.

This commit introduces a new object CMDQ that schedules an arbitrary
number of commands to the fixed-size physical command queue to
overcome the limitation. It replaces the HCA_factory and enforces a
better coding style for asynchronous commands.